refactor(delete): Restructure delete pkg into sub-packages
A similar structure as in pkg/deploy is now in place in pkg/delete.
Additionally versions of the All delete functions used for purge
are introduced, which do not collect errors into a slice for presentation
at the end of execution.
